--- app/validators/account_reset_cancel_token_validator.rb
@@ -1,16 +1,0 @@
-module AccountResetCancelTokenValidator
-  extend ActiveSupport::Concern
-
-  included do
-    validates :token, presence: true
-    validate :valid_token
-  end
-
-  private
-
-  def valid_token
-    return if account_reset_request
-
-    errors.add(:token, 'invalid')
-  end
-end
